launcher / finder changes
